Disable the ability to forcefully kill sidekiq process
Sidekiq supports `TERM` signal, such that, when received one,
it gracefully attempts to pause the workers, and shutsdown.

Hence, killing a process manually here, can be destructive, as
we are not allowing sidekiq to fullfil the worker completion cycle.
Which can also lead to unexpected behaviors of leaving jobs/batches
in private queues.

Moving forward, there won't be any `shutdown_signal` and
`shutdown_wait_timeout` config options. When a `TERM` signal is
received, sidekiq attempts to wait for the default `timeout` (`-t`)
period supplied at the time of sidekiq process boot. If a process,
does not terminate within the timeout period, its forcefully terminated.